Some refactoring to command handlers

+++ src/FMBot.Bot/Services/IndexService.cs
@@ -228,21 +228,17 @@ namespace FMBot.Bot.Services
                 .Include(i => i.GuildUsers)
                 .FirstOrDefaultAsync(f => f.DiscordGuildId == user.Guild.Id);
 
-            if (guild != null && guild.GuildUsers.Select(g => g.UserId).Contains(userThatLeft.UserId))
+            if (guild?.GuildUsers != null && guild.GuildUsers.Any() && guild.GuildUsers.Select(g => g.UserId).Contains(userThatLeft.UserId))
             {
-                var guildUser = guild.GuildUsers.FirstOrDefault(f => f.UserId == userThatLeft.UserId && f.GuildId == guild.GuildId);
+                var guildUser = guild
+                    .GuildUsers
+                    .FirstOrDefault(f => f.UserId == userThatLeft.UserId && f.GuildId == guild.GuildId);
 
                 if (guildUser != null)
                 {
                     db.GuildUsers.Remove(guildUser);
 
-                    Log.Information("Removed user {userId} from guild {guildName}", userThatLeft.UserId, guild.Name);
-
                     await db.SaveChangesAsync();
-                }
-                else
-                {
-                    Log.Warning("Tried removing user {userId} from guild {guildName}, but user was not stored in guild", userThatLeft.UserId, guild.Name);
                 }
             }
         }
